add: three "Translate to target language" buttons;
To multiple languages user, it is a very useful feature.
You can switch to the target language you want with only one click.
But I didn't figure out a clearer way to display these icon (now the icons are "A", "B", "C").
The icons will be update if I figure out a better way.