Remake the mechanism of playing audio; (The place play audio was background. Now is page.)
After remaking, audio can be paused,
the whole project could be upgraded to MV3 more smoothly.

But there might be some users playing audio and then close page.
They can't do it after the remaking, it is a compromise.

Also, the last audio will be cached now.
The experience of playing audio with the same text multiple times will be better.