fix: playing audio in website might violate CSP;
In some websites, play the audio with DataURL might violate the CSP(media-src).

The only way to get rid of it might be using "iframe".

I don't like this but I have to.

There might be some websites getting the message from the iframe and make some wrong reactions.

Will keep digging a better way to fix this issue.