refactor: in CSP violated website, use Web Speech API to play audio; remove: all relate to play audio in iframe;
Before, in CSP violated website, use iframe to play audio.
But there is a bug in MV3 that I have mentioned in #28.

After the refactor, use can't listen to audio of the language that the browser dosen't support in some websites which contain CSP(media-src '') like "github".

Closes #28

typofix: in "/src/public/translate/mojidict/translate.ts"