Revert the fixing of "Error: The message port closed before a response was received"
I find that it is not a good way to fix it, this kind of error happening
everywhere, and it is complex to fix it completely.

Fixing it completely will make this project hard to maintain later.

It shoulds be fixed by Chromium.

* Why it is a bug of Chromium?

`chrome.runtime.sendMessage` returns promise in MV3 in v99 if not pass a
callback, browser asume that the sender must want a response now.

It is abnormall.