Clicking "collect" button in collected state will not remove item from collection now
Instead, the item in collection will be updated

It seems it is not suitable that item in collection can be removed through translation panel.
So I decide to make it updatable and user can update it through clicking "collect" button even that is already collected.
And the date will be updated by clicking "collect" button every time.